# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "L5Swagger",
        "description": "Implementation of Swagger with in Laravel",
        "contact": {
            "email": "admin@admin.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost/api",
            "description": "Local"
        },
        {
            "url": "http://localhost/api/auth",
            "description": "Production"
        }
    ],
    "paths": {
...
        "/car": {
            "get": {
                "tags": [
                    "Машины"
                ],
                "summary": "Получить все ТС",
                "description": "Получить список ТС",
                "operationId": "2e8be4891000f3654c0c976fe7aa00ec",
                "responses": {
                    "200": {
                        "description": "Успех",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "oneOf": [
                                        {
                                            "$ref": "#/components/schemas/Car"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Не авторизован",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "oneOf": [
                                        {
                                            "$ref": "#/components/schemas/Response401"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Доступ запрещен",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "oneOf": [
                                        {
                                            "$ref": "#/components/schemas/Response403"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Не найдено",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "oneOf": [
                                        {
                                            "$ref": "#/components/schemas/Response404"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                }
            }
        },
        "/car/{id}": {
            "get": {
                "tags": [
                    "Машины"
                ],
                "summary": "Получить ТС",
                "description": "Получает ТС по идентификатору и возвращает его",
                "operationId": "a2cd332bc08309f5608d59b91e074e43",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Идентификатор пользователя",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "ca327b1a-ed73-41c6-afe0-1eca33866ec3"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успех",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "oneOf": [
                                        {
                                            "$ref": "#/components/schemas/Car"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Не авторизован",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "oneOf": [
                                        {
                                            "$ref": "#/components/schemas/Response401"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Доступ запрещен",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "oneOf": [
                                        {
                                            "$ref": "#/components/schemas/Response403"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Не найдено",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "oneOf": [
                                        {
                                            "$ref": "#/components/schemas/Response404"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                }
            }
        },
        "/car/create": {
            "post": {
                "tags": [
                    "Машины"
                ],
                "summary": "Создать ТС",
                "description": "Создает новое ТС и возвращает ее",
                "operationId": "62e6191f96dc48b2fec26e00908dc1be",
                "parameters": [
                    {
                        "name": "model_id",
                        "in": "path",
                        "description": "Идентификатор модели",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "0b4932f2-5c19-4de2-9ddc-17ce2375d164"
                        }
                    },
                    {
                        "name": "status",
                        "in": "path",
                        "description": "Статус аренды",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "rented"
                        }
                    },
                    {
                        "name": "mileage",
                        "in": "path",
                        "description": "Пробег ТС",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": "10383"
                        }
                    },
                    {
                        "name": "license_plate",
                        "in": "path",
                        "description": "Гос.номер ТС",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "J949YJ 93"
                        }
                    },
                    {
                        "name": "year",
                        "in": "path",
                        "description": "Год производства ТС",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": "2003"
                        }
                    },
                    {
                        "name": "location",
                        "in": "path",
                        "description": "Координаты текущего местоположения ТС",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "-35.71 -45.96609"
                        }
                    },
                    {
                        "name": "price_minute",
                        "in": "path",
                        "description": "Минутная цена аренды",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": "2"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "allOf": [
                                    {
                                        "$ref": "#/components/schemas/CarCreate"
                                    }
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Успех",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "oneOf": [
                                        {
                                            "$ref": "#/components/schemas/Car"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Не авторизован",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "oneOf": [
                                        {
                                            "$ref": "#/components/schemas/Response401"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Доступ запрещен",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "oneOf": [
                                        {
                                            "$ref": "#/components/schemas/Response403"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Не найдено",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "oneOf": [
                                        {
                                            "$ref": "#/components/schemas/Response404"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                }
            }
        },
        "/car/{id}/update": {
            "put": {
                "tags": [
                    "Машины"
                ],
                "summary": "Обновить ТС",
                "description": "Обновляет запись о ТС и возвращает ее",
                "operationId": "fdc909c26f47afaa130ed58b5517c212",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Существующий идентификатор ТС",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "ca327b1a-ed73-41c6-afe0-1eca33866ec3"
                        }
                    },
                    {
                        "name": "id",
                        "in": "query",
                        "description": "Идентификатор ТС",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "ca327b1a-ed73-41c6-afe0-1eca33866ec3"
                        }
                    },
                    {
                        "name": "model_id",
                        "in": "query",
                        "description": "Cчет по умолчанию",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "0b4932f2-5c19-4de2-9ddc-17ce2375d164"
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Статус аренды",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "rented"
                        }
                    },
                    {
                        "name": "mileage",
                        "in": "query",
                        "description": "Пробег ТС",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": "10383"
                        }
                    },
                    {
                        "name": "license_plate",
                        "in": "query",
                        "description": "Гос.номер ТС",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "Н709ОM 147"
                        }
                    },
                    {
                        "name": "year",
                        "in": "query",
                        "description": "Год производства ТС",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": "2003"
                        }
                    },
                    {
                        "name": "location",
                        "in": "query",
                        "description": "Координаты текущего местоположения ТС",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "-35.71 -45.96609"
                        }
                    },
                    {
                        "name": "price_minute",
                        "in": "query",
                        "description": "Минутная цена аренды",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": "2"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "allOf": [
                                    {
                                        "$ref": "#/components/schemas/Car"
                                    }
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Успех",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "oneOf": [
                                        {
                                            "$ref": "#/components/schemas/Car"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Не авторизован",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "oneOf": [
                                        {
                                            "$ref": "#/components/schemas/Response401"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Доступ запрещен",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "oneOf": [
                                        {
                                            "$ref": "#/components/schemas/Response403"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Не найдено",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "oneOf": [
                                        {
                                            "$ref": "#/components/schemas/Response404"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Car": {
                "title": "Car",
                "properties": {
                    "id": {
                        "type": "uuid",
                        "example": "ca327b1a-ed73-41c6-afe0-1eca33866ec3"
                    },
                    "model_id": {
                        "type": "uuid",
                        "example": "0b4932f2-5c19-4de2-9ddc-17ce2375d164"
                    },
                    "status": {
                        "type": "string",
                        "example": "rented"
                    },
                    "mileage": {
                        "type": "integer",
                        "example": 10383
                    },
                    "license_plate": {
                        "type": "string",
                        "example": "J949YJ 93"
                    },
                    "year": {
                        "type": "integer",
                        "example": 2003
                    },
                    "location": {
                        "type": "string",
                        "example": "-35.71 -45.96609"
                    },
                    "price_minute": {
                        "type": "integer",
                        "example": 2
                    }
                },
                "type": "object"
            },
...
            "Response401": {
                "title": "Response401",
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Unauthorized"
                    }
                },
                "type": "object"
            },
            "Response403": {
                "title": "Response403",
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Forbidden"
                    }
                },
                "type": "object"
            },
            "Response404": {
                "title": "Response404",
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Not Found"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}